+++ src/Bio/ABI/Type.hs
@@ -1,17 +1,17 @@
 module Bio.ABI.Type
-  ( ABIProcessed (..)
+  ( ABIBasecalled (..)
   ) where
 
 import           Bio.Sequence (IsSequence (..), WeightedSequence)
 import           Data.Coerce  (coerce)
 
-newtype ABIProcessed = ABIProcessed (WeightedSequence Double Char)
+newtype ABIBasecalled = ABIBasecalled (WeightedSequence Double Char)
   deriving (Eq, Show)
 
-instance IsSequence ABIProcessed where
-  type Element ABIProcessed = Char
-  type Marking ABIProcessed = ()
-  type Weight ABIProcessed  = Double
+instance IsSequence ABIBasecalled where
+  type Element ABIBasecalled = Char
+  type Marking ABIBasecalled = ()
+  type Weight ABIBasecalled  = Double
 
   toSequence = coerce
   fromSequence = coerce